Fixed brackets and indented code
The first column in the 'myhtmlbodymacro'-template was missing a bracket:
  {book.tag}} --> {{book.tag}}

The code has been indented for readibility.

Moreover, a condition regarding the entity's properties length has been added to the submit-row. This was necessary because of the (hidden) submit button.

The StationTrack-issue from submit (edae03b46d5123fb30aba) has been solved -
StationTrack resulted from a different generation-process and therefore
was not up to date.