Templates use viewRef-information
This commit extends the view-templates:

The mainView now generates a html-document, which includes html- and
js-documents belonging to the referenced views (viewRef) and a js-file
called 'app.js'.
For each view a html-file and a js-file are generated. The html-file
contains some information (typically the view's entity), while the js-file
contains angular-code (a controller for interaction and directives).
These js- and html-files are automatically included by the mainView's
html-file. Additionally the view's controllers are included in the
'app.js'-file via dependency injection (DI).

Example folder-content:
- FirstView.html
- FirstView.js
- SndView.html
- SndView.js

- index.html contains
 --> <script src="***View.js">
 --> <section ng-include="***View.html">

- app.js contains DI
 --> angular.module("MainView",["FirstView","SecondView"])

- angular.js + boostrap.css