Implemented table-layout-module
This commit changes the structure of commit 1ae9be98dbb5b85def slightly.
The main change is the ControllerBase.js-file; it acts like a superclass
for different templates (e.g. TableControllerBase). It is included in the
index.html and used by the viewRefs' js-files.

Example folder-content:

- FirstView.js + SndView.js
 --> 'Extend' ControllerBase.js

- index.html contains
 --> <script src="TableControllerBase.js"> //For a table-layout

- FirstView.html
- SndView.html

- app.js
- angular.js + boostrap.css

Because of angular's controller registration, the extension of
ControllerBase.js is neither a class-inheritance nor a
prototype-based-inheritance.

The problem was, that angular registers controllers via a method, which
expects a string and a function, as opposed to our planning, in which we
tried to pass an object to the function.
The solution to this is an injected function; when the injected function
is executed (with the injector as a parameter), it returns a functions,
which's parameter is $scope. When this is executed, the injector adds the
'extended' methods to the $scope.