Add button-lightbox functionality (add/delete)
A service has been added, that allows for a lightbox to pop up when a
button is pushed. At the moment, the lightbox contains mockup-data and
an interface for adding and deleting an entry.

Next, a communicaton will be established between the view-controller and
table-controller. This allows for passing data to the lightbox:
 - which Task is associated with the pressed button (column structure)
 - which data is associated with the selected element (show data before
deletion)